Conflicts:
	openwis-dataservice/openwis-dataservice-cache/openwis-dataservice-cache-core/pom.xml
	openwis-dataservice/openwis-dataservice-cache/openwis-dataservice-cache-ejb/pom.xml
	openwis-dataservice/openwis-dataservice-cache/openwis-dataservice-cache-ftpreplication/pom.xml
	openwis-dataservice/openwis-dataservice-cache/openwis-dataservice-cache-webapp/pom.xml
	openwis-dataservice/openwis-dataservice-cache/pom.xml
	openwis-dataservice/openwis-dataservice-common/openwis-dataservice-common-domain/pom.xml
	openwis-dataservice/openwis-dataservice-common/openwis-dataservice-common-timer/pom.xml
	openwis-dataservice/openwis-dataservice-common/openwis-dataservice-common-utils/pom.xml
	openwis-dataservice/openwis-dataservice-common/pom.xml
	openwis-dataservice/openwis-dataservice-server/openwis-dataservice-server-ear/pom.xml
	openwis-dataservice/openwis-dataservice-server/openwis-dataservice-server-ejb/pom.xml
	openwis-dataservice/openwis-dataservice-server/openwis-dataservice-server-webapp/pom.xml
	openwis-dataservice/openwis-dataservice-server/pom.xml
	openwis-dataservice/pom.xml
	openwis-harness/openwis-harness-client/pom.xml
	openwis-harness/openwis-harness-dissemination/pom.xml
	openwis-harness/openwis-harness-localdatasource/pom.xml
	openwis-harness/openwis-harness-mssfss/pom.xml
	openwis-harness/openwis-harness-subselectionparameters/pom.xml
	openwis-harness/pom.xml
	openwis-management/openwis-management-client/pom.xml
	openwis-management/openwis-management-service/openwis-management-service-common/pom.xml
	openwis-management/openwis-management-service/openwis-management-service-ear/pom.xml
	openwis-management/openwis-management-service/openwis-management-service-ejb/pom.xml
	openwis-management/openwis-management-service/pom.xml
	openwis-management/pom.xml
	openwis-metadataportal/cachingxslt/pom.xml
	openwis-metadataportal/jeeves/pom.xml
	openwis-metadataportal/oaipmh/pom.xml
	openwis-metadataportal/openwis-portal-solr/pom.xml
	openwis-metadataportal/openwis-portal/pom.xml
	openwis-metadataportal/openwis-portal/src/main/java/org/openwis/metadataportal/kernel/metadata/product/ProductMetadataDefaultExtractor.java
	openwis-metadataportal/openwis-portal/src/main/java/org/openwis/metadataportal/services/register/SelfRegister.java
	openwis-metadataportal/pom.xml
	openwis-metadataportal/sde/pom.xml
	openwis-portal-client/pom.xml
	openwis-securityservice/openwis-securityservice-usermanagement-common/pom.xml
	openwis-securityservice/openwis-securityservice-usermanagement-server/openwis-securityservice-usermanagement-server-ear/pom.xml
	openwis-securityservice/openwis-securityservice-usermanagement-server/openwis-securityservice-usermanagement-server-ear/src/main/config/openwis-securityservice-jndi-service.xml
	openwis-securityservice/openwis-securityservice-usermanagement-server/openwis-securityservice-usermanagement-server-ejb/pom.xml
	openwis-securityservice/openwis-securityservice-usermanagement-server/openwis-securityservice-usermanagement-server-webapp/pom.xml
	openwis-securityservice/openwis-securityservice-usermanagement-server/pom.xml
	openwis-securityservice/openwis-securityservice-utils/PopulateLDAP/pom.xml
	openwis-securityservice/openwis-securityservice-utils/pom.xml
	openwis-securityservice/pom.xml
	openwis-stagingpost/pom.xml
	pom.xml

+++ openwis-dataservice/openwis-dataservice-server/openwis-dataservice-server-ejb/src/main/java/org/openwis/datasource/server/service/impl/BlacklistServiceImpl.java
@@ -308,11 +308,11 @@ public class BlacklistServiceImpl implements BlacklistService {
          result.setNbDisseminationWarnThreshold(JndiUtils
                .getLong(DataServiceConfiguration.BLACKLIST_DEFAULT_NB_WARN));
          result.setNbDisseminationBlacklistThreshold(JndiUtils
-               .getInt(DataServiceConfiguration.BLACKLIST_DEFAULT_NB_BLACKLIST));
+               .getLong(DataServiceConfiguration.BLACKLIST_DEFAULT_NB_BLACKLIST));
          result.setVolDisseminationWarnThreshold(JndiUtils
-               .getInt(DataServiceConfiguration.BLACKLIST_DEFAULT_VOL_WARN));
+               .getLong(DataServiceConfiguration.BLACKLIST_DEFAULT_VOL_WARN));
          result.setVolDisseminationBlacklistThreshold(JndiUtils
-               .getInt(DataServiceConfiguration.BLACKLIST_DEFAULT_VOL_BLACKLIST));
+               .getLong(DataServiceConfiguration.BLACKLIST_DEFAULT_VOL_BLACKLIST));
 
       }
       return result;
